# openapi/state-v1.yaml
openapi: 3.1.0
info:
  title: State API
  version: 1.0.0
  description: REST contract for the self-hosted State reminder and context service.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://state.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/notes:
    post:
      tags: [Notes]
      operationId: createNote
      description: Owner, device and harness. Stores one note. Without a title the first line of the document becomes the title (title_source derived).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note stored with its audit event.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    get:
      tags: [Notes]
      operationId: listNotes
      description: Owner, device and harness. Newest change first. Runners receive 403, and note events are removed from the change feed and briefing they read.
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Full-text query over title, summary and plain text. }
        - { name: include_archived, in: query, schema: { type: boolean, default: false } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Matching notes.
          content:
            application/json:
              schema:
                type: object
                required: [notes]
                properties:
                  notes:
                    type: array
                    items: { $ref: "#/components/schemas/NoteView" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}:
    parameters:
      - $ref: "#/components/parameters/NoteID"
    get:
      tags: [Notes]
      operationId: getNote
      description: Owner, device and harness read access.
      responses:
        "200":
          description: Note.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Notes]
      operationId: updateNote
      description: Optimistic revision checking. Archiving and restoring (archived) is limited to the owner and devices; a stale revision answers 409 with details.server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Updated note.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}/history:
    parameters:
      - $ref: "#/components/parameters/NoteID"
    get:
      tags: [Notes]
      operationId: getNoteHistory
      description: Complete audit history of one note.
      responses:
        "200":
          description: Audit events, oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEvent" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    NoteID:
      name: note_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUIDv7" }
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    UUIDv7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
...
    ActorKind:
      type: string
      enum: [owner, device, harness, system, runner]
    Actor:
      type: object
      required: [id, kind]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        kind: { $ref: "#/components/schemas/ActorKind" }
        display_name: { type: [string, "null"] }
        harness: { type: [string, "null"] }
        device_name: { type: [string, "null"] }
...
    AuditEvent:
      type: object
      required: [id, reminder_id, action, actor, server_time, changed_fields, revision, correlation_id, client_request_id, hash, signature]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        reminder_id:
          type: string
          description: UUIDv7 of the reminder, or an empty string for project, policy, runner and note events.
        note_id:
          $ref: "#/components/schemas/UUIDv7"
          description: Present on note events (note.created, note.updated, note.archived, note.restored), which carry no reminder.
        action: { type: string }
        actor: { $ref: "#/components/schemas/Actor" }
        server_time: { type: string, format: date-time }
        client_time: { type: [string, "null"], format: date-time }
        source: { type: [string, "null"] }
        source_excerpt: { type: [string, "null"] }
        before_snapshot: { type: [object, array, string, number, boolean, "null"] }
        after_snapshot: { type: [object, array, string, number, boolean, "null"] }
        changed_fields:
          type: array
          items: { type: string }
        revision: { type: integer, format: int64, minimum: 1 }
        correlation_id: { type: string }
        client_request_id: { type: string }
        previous_hash: { type: [string, "null"] }
        hash: { type: string }
        signature: { type: string }
...
    MutationMetadata:
      type: object
      required: [client_request_id]
      properties:
        client_time: { type: [string, "null"], format: date-time }
        source: { type: string, maxLength: 200 }
        source_excerpt: { type: [string, "null"], maxLength: 10000 }
        client_request_id: { type: string, minLength: 1, maxLength: 200 }
        correlation_id: { type: string, maxLength: 200 }
...
    NoteFieldSource:
      type: string
      enum: [user, derived, ai]
      description: user values are never overwritten automatically; derived values follow the document; ai values come from the notes AI and never replace a user value.
    Note:
      type: object
      required: [id, title, title_source, document, plain_text, summary, summary_source, archived, revision, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        title: { type: string, maxLength: 200 }
        title_source: { $ref: "#/components/schemas/NoteFieldSource" }
        document:
          type: string
          description: Canonical content as Markdown (CommonMark plus - [ ] checklists and --- dividers). At most 262144 UTF-8 bytes; JSON Schema maxLength counts characters, so the byte limit is enforced by the server.
        plain_text: { type: string, description: Derived text without Markdown markers, used for search. }
        summary: { type: string, maxLength: 280 }
        summary_source: { $ref: "#/components/schemas/NoteFieldSource" }
        capture: { type: string, enum: [image, audio], description: Set for photo and voice notes, which may start without text. }
        archived: { type: boolean }
        revision: { type: integer, format: int64, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NoteView:
      description: A note as clients see it. An AI title or summary replaces only a derived one.
      allOf:
        - $ref: "#/components/schemas/Note"
        - type: object
          required: [attachments, processing, relations, reminder_proposals]
          properties:
            attachments:
              type: array
              items: { $ref: "#/components/schemas/NoteAttachment" }
            processing: { $ref: "#/components/schemas/NoteProcessing" }
            ai: { $ref: "#/components/schemas/NoteAIResult" }
            relations:
              type: array
              items: { $ref: "#/components/schemas/NoteRelation" }
            reminder_proposals:
              type: array
              items: { $ref: "#/components/schemas/ReminderProposal" }
    NoteAttachment:
      type: object
      required: [id, note_id, ordinal, kind, mime_type, byte_size, sha256, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        ordinal: { type: integer, minimum: 0 }
        kind: { type: string, enum: [image, audio] }
        mime_type: { type: string }
        byte_size: { type: integer, format: int64 }
        sha256: { type: string }
        duration_ms: { type: integer, format: int64 }
        derived_text: { type: string, description: OCR text of a photo or transcript of a recording. }
        derived_kind: { type: string, enum: [ocr, transcript] }
        derived_model: { type: string }
        created_at: { type: string, format: date-time }
    NoteProcessing:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [idle, queued, processing, ready, needs_review, failed, consent_required, not_configured, budget_exhausted] }
        error: { type: string }
        model: { type: string }
        updated_at: { type: string, format: date-time }
    NoteAIResult:
      type: object
      properties:
        title: { type: string }
        summary: { type: string }
        model: { type: string }
        source_hash: { type: string, description: SHA-256 of the plain text the result was made from. }
        proposed_document: { type: string, description: Body the AI suggests when the owner already wrote text. }
        needs_review: { type: boolean }
        updated_at: { type: string, format: date-time }
    NoteRelation:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        related_note_id: { $ref: "#/components/schemas/UUIDv7" }
        related_title: { type: string }
        reason: { type: string }
        confidence: { type: number }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    ReminderProposal:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        title: { type: string }
        description: { type: string }
        local_date: { type: string }
        local_time: { type: string }
        reason: { type: string }
        status: { type: string, enum: [pending, accepted, dismissed] }
        reminder_id: { type: string }
...
    CreateNoteRequest:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          description: A title or a non-empty document is required, except for a photo or voice note (capture image or audio).
          properties:
            capture: { type: string, enum: [text, image, audio] }
            title: { type: string, maxLength: 200 }
            document: { type: string, description: At most 262144 UTF-8 bytes. }
            summary: { type: string, maxLength: 280 }
    UpdateNoteRequest:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [expected_revision]
          properties:
            title: { type: string, maxLength: 200, description: An empty string returns to the derived title. }
            document: { type: string, description: At most 262144 UTF-8 bytes. }
            summary: { type: string, maxLength: 280, description: An empty string returns to the derived summary. }
            archived: { type: boolean, description: Owner and devices only. }
            expected_revision: { type: integer, format: int64, minimum: 1 }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: [object, array, string, number, boolean, "null"] }